add additional system information to the debug filter

+++ src/net/instaweb/rewriter/rewrite_context_test.cc
@@ -1717,7 +1717,8 @@ TEST_F(RewriteContextTest, CacheExtendCacheableResource) {
                               &headers));
     EXPECT_EQ("a", content);
     EXPECT_STREQ(StringPrintf("max-age=%lld",
-                              ServerContext::kGeneratedMaxAgeMs/1000),
+                              static_cast<long long int>(
+                                  ServerContext::kGeneratedMaxAgeMs / 1000)),
                  headers.Lookup1(HttpAttributes::kCacheControl));
   }
 }
